+++ docs/src/examples/templates/conditionals-if/component.js
@@ -1,0 +1,15 @@
+import { defineComponent, getText } from '@semantic-ui/component';
+
+const css = await getText('./component.css');
+const template = await getText('./component.html');
+
+const defaultState = {
+  isActive: true,
+};
+
+export default defineComponent({
+  tagName: 'if-condition',
+  template,
+  css,
+  defaultState,
+});